try to bump aap to 0.9.0 and various other deps.
The app now fails to launch saying that it cannot find libc++ symbols, and
it is indeed likely to happen because aap-guitarix does not build anything C++
within the app itself. This does not get bundled even if we import
libcxx-provider (because the app does not use anything from the aar after all).

We should probably retire aap-lv2-guitarix and use guitarix.vst project
instead (which is a JUCE project nowaday).